requirements.txt
slack-sdk==3.21.3
python-dotenv==1.0.0
elasticsearch==8.8.0
requests==2.31.0
configparser==5.3.0
logging==0.4.9.6